ab_test_platform/requirements.txt
fastapi==0.129.0
sqlalchemy==2.0.46
sqlmodel==0.0.33
asyncpg==0.31.0
uvicorn==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
PyJWT==2.11.0
python-dotenv==1.2.1
lark==1.3.1
email-validator==2.3.0
tavern[pytest]
pytest-cov
